requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
h5py>=3.9.0
opencv-python>=4.8.0
scipy>=1.11.0
tqdm>=4.66.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
tensorboard>=2.14.0
pyrender>=0.1.45
trimesh>=3.23.5
chumpy>=0.70
smplx>=0.1.28
pybullet>=3.2.5
ultralytics>=8.0.0
psutil>=5.9.0
openpyxl>=3.1.0
google-colab>=1.0.0
ipython>=8.0.0